+++ Demo/Application/Features/3D Secure/BraintreeDemoThreeDSecurePaymentFlowViewController.h
@@ -2,8 +2,6 @@
 
 #import "BraintreeDemoPaymentButtonBaseViewController.h"
 
-@class BTThreeDSecureClient;
-
 @interface BraintreeDemoThreeDSecurePaymentFlowViewController : BraintreeDemoPaymentButtonBaseViewController
 
 @end
