+++ Sources/BraintreeThreeDSecure/BTThreeDSecureRequest.swift
@@ -78,246 +78,17 @@ import BraintreePaymentFlow
     var dfReferenceID: String? = nil
 
     var threeDSecureV2Provider: BTThreeDSecureV2Provider?
-    
-    // MARK: - Internal Methods
-    
-    /// Prepare for a 3DS 2.0 flow.
-    /// - Parameters:
-    ///   - apiClient: The API client.
-    ///   - completion: This completion will be invoked exactly once. If the error is nil then the preparation was successful.
-    func prepareLookup(
-        apiClient: BTAPIClient,
-        completion: @escaping (Error?) -> Void
-    ) {
-        apiClient.fetchOrReturnRemoteConfiguration { [weak self] configuration, error in
-            guard let self else { return }
-
-            guard let configuration, error == nil else {
-                completion(error)
-                return
-            }
-
-            if configuration.cardinalAuthenticationJWT != nil {
-                self.threeDSecureV2Provider = BTThreeDSecureV2Provider(
-                    configuration: configuration,
-                    apiClient: apiClient,
-                    request: self
-                ) { lookupParameters in
-                    if let dfReferenceID = lookupParameters?["dfReferenceId"] {
-                        self.dfReferenceID = dfReferenceID
-                    }
-                    completion(nil)
-                }
-            } else {
-                completion(BTThreeDSecureError.configuration("Merchant is not configured for 3SD 2."))
-                return
-            }
-        }
-    }
-
-    func process(lookupResult: BTThreeDSecureResult, configuration: BTConfiguration) {
-        if lookupResult.lookup?.requiresUserAuthentication == false || lookupResult.lookup == nil {
-            paymentFlowClientDelegate?.onPaymentComplete(lookupResult, error: nil)
-            return
-        }
-
-        if lookupResult.lookup?.isThreeDSecureVersion2 == true {
-            performV2Authentication(with: lookupResult)
-        }
-    }
-    
-    // MARK: - Private Methods
-    
-    private func start(request: BTPaymentFlowRequest, configuration: BTConfiguration) {
-        guard let apiClient = paymentFlowClientDelegate?.apiClient() else {
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: BTThreeDSecureError.invalidAPIClient)
-            return
-        }
-
-        guard let threeDSecureRequest = request as? BTThreeDSecureRequest else {
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: BTThreeDSecureError.cannotCastRequest)
-            return
-        }
-
-        let paymentFlowClient = BTPaymentFlowClient(apiClient: apiClient)
-
-        if threeDSecureRequest.threeDSecureRequestDelegate == nil {
-            threeDSecureRequest.threeDSecureRequestDelegate = self
-        }
-
-        apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.started")
-        paymentFlowClient.performThreeDSecureLookup(threeDSecureRequest) { lookupResult, error in
-            DispatchQueue.main.async {
-                guard let lookupResult, error == nil else {
-                    apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.failed")
-                    self.paymentFlowClientDelegate?.onPayment(with: nil, error: error)
-                    return
-                }
-
-                let threeDSecureVersion = lookupResult.lookup?.threeDSecureVersion ?? "2"
-                apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.3ds-version.\(threeDSecureVersion)")
-
-                self.threeDSecureRequestDelegate?.onLookupComplete(threeDSecureRequest, lookupResult: lookupResult) {
-                    let requiresUserAuthentication = lookupResult.lookup?.requiresUserAuthentication ?? false
-                    apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.challenge-presented.\(self.stringFor(requiresUserAuthentication))")
-                    self.process(lookupResult: lookupResult, configuration: configuration)
-                }
-            }
-        }
-    }
-    
-    private func performV2Authentication(with lookupResult: BTThreeDSecureResult) {
-        guard let apiClient = paymentFlowClientDelegate?.apiClient() else {
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: BTThreeDSecureError.invalidAPIClient)
-            return
-        }
-
-        threeDSecureV2Provider?.process(lookupResult: lookupResult) { result, error in
-            guard let result else {
-                apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.failed")
-                self.paymentFlowClientDelegate?.onPaymentComplete(nil, error: error)
-                return
-            }
-
-            self.logThreeDSecureCompletedAnalytics(forResult: lookupResult, apiClient: apiClient)
-            self.paymentFlowClientDelegate?.onPaymentComplete(result, error: error)
-        }
-    }
-    
-    private func logThreeDSecureCompletedAnalytics(forResult result: BTThreeDSecureResult, apiClient: BTAPIClient) {
-        let liabilityShiftPossible = result.tokenizedCard?.threeDSecureInfo.liabilityShiftPossible ?? false
-        apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.liability-shift-possible.\(stringFor(liabilityShiftPossible))")
-
-        let liabilityShifted = result.tokenizedCard?.threeDSecureInfo.liabilityShiftPossible ?? false
-        apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.liability-shifted.\(liabilityShifted)")
-
-        apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.completed")
-    }
-    
-    func stringFor(_ boolean: Bool) -> String {
-        boolean ? "true" : "false"
-    }
-}
-
-// MARK: - BTPaymentFlowRequestDelegate Protocol Conformance
-
-extension BTThreeDSecureRequest: BTPaymentFlowRequestDelegate {
-
-    /// :nodoc:
-    public func handle(
-        _ request: BTPaymentFlowRequest,
-        client apiClient: BTAPIClient,
-        paymentClientDelegate delegate: BTPaymentFlowClientDelegate
-    ) {
-        paymentFlowClientDelegate = delegate
-
-        apiClient.sendAnalyticsEvent("ios.three-d-secure.initialized")
-
-        apiClient.fetchOrReturnRemoteConfiguration { [weak self] configuration, error in
-            guard let self else { return }
-
-            if let error {
-                self.paymentFlowClientDelegate?.onPaymentComplete(nil, error: error)
-                return
-            }
-
-            var integrationError: Error?
-
-            if configuration?.cardinalAuthenticationJWT == nil {
-                NSLog("%@ BTThreeDSecureRequest versionRequested is 2, but merchant account is not setup properly.", BTLogLevelDescription.string(for: .critical))
-                integrationError = BTThreeDSecureError.configuration("BTThreeDSecureRequest versionRequested is 2, but merchant account is not setup properly.")
-            }
-
-            if self.amount?.decimalValue.isNaN == true || self.amount == nil {
-                NSLog("%@ BTThreeDSecureRequest amount can not be nil or NaN.", BTLogLevelDescription.string(for: .critical))
-                integrationError = BTThreeDSecureError.configuration("BTThreeDSecureRequest amount can not be nil or NaN.")
-            }
-
-            if self.threeDSecureRequestDelegate == nil {
-                integrationError = BTThreeDSecureError.configuration("Configuration Error: threeDSecureRequestDelegate can not be nil when versionRequested is 2.")
-            }
-
-            if let integrationError {
-                delegate.onPaymentComplete(nil, error: integrationError)
-                return
-            }
-
-            guard let configuration, configuration.cardinalAuthenticationJWT != nil else {
-                delegate.onPaymentComplete(nil, error: BTThreeDSecureError.configuration("Merchant does not have the required Cardinal authentication JWT."))
-                return
-            }
-
-            self.prepareLookup(apiClient: apiClient) { error in
-                if let error {
-                    delegate.onPaymentComplete(nil, error: error)
-                    return
-                }
-
-                self.start(request: request, configuration: configuration)
-            }
-        }
-    }
-
-    /// :nodoc:
-    public func handleOpen(_ url: URL) {
-        guard let jsonAuthResponse: String = BTURLUtils.queryParameters(for: url)["auth_response"],
-                jsonAuthResponse.count != 0 else {
-            paymentFlowClientDelegate?.apiClient().sendAnalyticsEvent("ios.three-d-secure.missing-auth-response")
-
-            let error = BTThreeDSecureError.failedAuthentication("Auth Response missing from URL.")
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: error)
-            return
-        }
-
-        guard let jsonAuthResponseData = jsonAuthResponse.data(using: .utf8) else {
-            paymentFlowClientDelegate?.apiClient().sendAnalyticsEvent("ios.three-d-secure.invalid-auth-data")
-
-            let error = BTThreeDSecureError.failedAuthentication("Auth Response cannot be converted to Data type.")
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: error)
-            return
-        }
-
-        guard let jsonData = try? JSONSerialization.jsonObject(with: jsonAuthResponseData) else {
-            paymentFlowClientDelegate?.apiClient().sendAnalyticsEvent("ios.three-d-secure.invalid-auth-response")
-
-            let error = BTThreeDSecureError.failedAuthentication("Auth Response JSON parsing error.")
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: error)
-            return
-        }
-
-        let authBody = BTJSON(value: jsonData)
-        let result = BTThreeDSecureResult(json: authBody)
-
-        guard let apiClient = paymentFlowClientDelegate?.apiClient() else {
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: BTThreeDSecureError.invalidAPIClient)
-            return
-        }
-
-        if let errorMessage = result.errorMessage, result.tokenizedCard == nil {
-            apiClient.sendAnalyticsEvent("ios.three-d-secure.verification-flow.failed")
-            paymentFlowClientDelegate?.onPaymentComplete(nil, error: BTThreeDSecureError.failedAuthentication(errorMessage))
-            return
-        }
-
-        logThreeDSecureCompletedAnalytics(forResult: result, apiClient: apiClient)
-        paymentFlowClientDelegate?.onPaymentComplete(result, error: nil)
-    }
-
-    /// :nodoc:
-    public func paymentFlowName() -> String {
-        "three-d-secure"
-    }
 }
 
 // MARK: - BTThreeDSecureRequestDelegate Protocol Conformance
 
-extension BTThreeDSecureRequest: BTThreeDSecureRequestDelegate {
-
-    public func onLookupComplete(
-        _ request: BTThreeDSecureRequest,
-        lookupResult result: BTThreeDSecureResult,
-        next: (() -> Void)
-    ) {
-        next()
-    }
-}
+//extension BTThreeDSecureRequest: BTThreeDSecureRequestDelegate {
+//
+//    public func onLookupComplete(
+//        _ request: BTThreeDSecureRequest,
+//        lookupResult result: BTThreeDSecureResult,
+//        next: (() -> Void)
+//    ) {
+//        next()
+//    }
+//}
