+++ UnitTests/BraintreeThreeDSecureTests/BTThreeDSecure_Tests.swift
@@ -4,6 +4,7 @@ import XCTest
 @testable import BraintreeCore
 @testable import BraintreeThreeDSecure
 
+// TODO: - Move to ThreeDSecureClient test file & organize with sut & method name
 class BTThreeDSecure_UnitTests: XCTestCase {
     let tempClientToken = "eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJmNTI0M2RkZGRmNzlkNGFiYmI5YjYwMDUzN2ZkZjQ0ZDViNDg0ODVkOWU0ZjJmYmI3YWM5ZTU2MGE3ZDVhZmM5fGNyZWF0ZWRfYXQ9MjAxNy0wNC0xM1QyMTozOTo0My40MjM4NzE4MTUrMDAwMFx1MDAyNmN1c3RvbWVyX2lkPTJENzJCNjQ4LUI0RkMtNDQ1My1BOURDLTI2QTYyMEVGNjQwNFx1MDAyNm1lcmNoYW50X2FjY291bnRfaWQ9aWRlYWxfZXVyXHUwMDI2bWVyY2hhbnRfaWQ9ZGNwc3B5MmJyd2RqcjNxblx1MDAyNnB1YmxpY19rZXk9OXd3cnpxazN2cjN0NG5jOCIsImNvbmZpZ1VybCI6Imh0dHBzOi8vYXBpLnNhbmRib3guYnJhaW50cmVlZ2F0ZXdheS5jb206NDQzL21lcmNoYW50cy9kY3BzcHkyYnJ3ZGpyM3FuL2NsaWVudF9hcGkvdjEvY29uZmlndXJhdGlvbiIsImNoYWxsZW5nZXMiOlsiY3Z2IiwicG9zdGFsX2NvZGUiXSwiZW52aXJvbm1lbnQiOiJzYW5kYm94IiwiY2xpZW50QXBpVXJsIjoiaHR0cHM6Ly9hcGkuc2FuZGJveC5icmFpbnRyZWVnYXRld2F5LmNvbTo0NDMvbWVyY2hhbnRzL2RjcHNweTJicndkanIzcW4vY2xpZW50X2FwaSIsImFzc2V0c1VybCI6Imh0dHBzOi8vYXNzZXRzLmJyYWludHJlZWdhdGV3YXkuY29tIiwiYXV0aFVybCI6Imh0dHBzOi8vYXV0aC52ZW5tby5zYW5kYm94LmJyYWludHJlZWdhdGV3YXkuY29tIiwiYW5hbHl0aWNzIjp7InVybCI6Imh0dHBzOi8vY2xpZW50LWFuYWx5dGljcy5zYW5kYm94LmJyYWludHJlZWdhdGV3YXkuY29tL2RjcHNweTJicndkanIzcW4ifSwidGhyZWVEU2VjdXJlRW5hYmxlZCI6ZmFsc2UsInBheXBhbEVuYWJsZWQiOmZhbHNlLCJjb2luYmFzZUVuYWJsZWQiOnRydWUsImNvaW5iYXNlIjp7ImNsaWVudElkIjoiN2U5NWUwZmRkYTE0ODQ2NjU4YjM4Zjc3MmJhMmQzMGNkNzhhOWYyMTQ0YzUzOTA4NmU1NzkwYmYzNzdmYmVlZCIsIm1lcmNoYW50QWNjb3VudCI6ImNvaW5iYXNlLXNhbmRib3gtc2hhcmVkLW1lcmNoYW50QGdldGJyYWludHJlZS5jb20iLCJzY29wZXMiOiJhdXRob3JpemF0aW9uczpicmFpbnRyZWUgdXNlciIsInJlZGlyZWN0VXJsIjoiaHR0cHM6Ly9hc3NldHMuYnJhaW50cmVlZ2F0ZXdheS5jb20vY29pbmJhc2Uvb2F1dGgvcmVkaXJlY3QtbGFuZGluZy5odG1sIiwiZW52aXJvbm1lbnQiOiJwcm9kdWN0aW9uIn0sImJyYWludHJlZV9hcGkiOnsiYWNjZXNzX3Rva2VuIjoic2FuZGJveF9mN2RyNWNfZHE2c3MyX2prczd4dF80aHNwc2hfcWI3IiwidXJsIjoiaHR0cHM6Ly9wYXltZW50cy5zYW5kYm94LmJyYWludHJlZS1hcGkuY29tIn0sIm1lcmNoYW50SWQiOiJkY3BzcHkyYnJ3ZGpyM3FuIiwidmVubW8iOiJvZmZsaW5lIiwiYXBwbGVQYXkiOnsic3RhdHVzIjoibW9jayIsImNvdW50cnlDb2RlIjoiVVMiLCJjdXJyZW5jeUNvZGUiOiJFVVIiLCJtZXJjaGFudElkZW50aWZpZXIiOiJtZXJjaGFudC5jb20uYnJhaW50cmVlcGF5bWVudHMuc2FuZGJveC5CcmFpbnRyZWUtRGVtbyIsInN1cHBvcnRlZE5ldHdvcmtzIjpbInZpc2EiLCJtYXN0ZXJjYXJkIiwiYW1leCIsImRpc2NvdmVyIl19LCJtZXJjaGFudEFjY291bnRJZCI6ImlkZWFsX2V1ciJ9"
     var mockAPIClient : MockAPIClient!
@@ -31,7 +32,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": [],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
 
         client.startPaymentFlow(threeDSecureRequest) { result, error in
             XCTAssertNotNil(error)
@@ -56,7 +57,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
 
         client.startPaymentFlow(threeDSecureRequest) { result, error in
             XCTAssertNotNil(error)
@@ -69,7 +70,8 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
         waitForExpectations(timeout: 4)
     }
-    
+
+    // TODO - duplicate test
     func testStartPayment_returnsError_whenAmountIsNotANumber() {
         threeDSecureRequest = BTThreeDSecureRequest()
         threeDSecureRequest.amount = .notANumber
@@ -82,7 +84,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
 
         client.startPaymentFlow(threeDSecureRequest) { result, error in
             XCTAssertNotNil(error)
@@ -96,6 +98,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 4)
     }
 
+    // TODO: - is this relevant?
     func testStartPayment_v2_doesNotDisplaySafariViewControllerWhenAuthenticationNotRequired() {
         threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
 
@@ -105,7 +108,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
         let responseBody = [
             "paymentMethod": [
                 "consumed": false,
@@ -147,6 +150,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 4)
     }
 
+    // TODO: - this test doesn't do what it says it does
     func testStartPayment_v2_callsOnLookupCompleteDelegateMethod() {
         threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
 
@@ -156,7 +160,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
 
         client.startPaymentFlow(threeDSecureRequest) { result, error in
             expectation.fulfill()
@@ -172,7 +176,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
 
         client.startPaymentFlow(threeDSecureRequest) { result, error in
             XCTAssertNotNil(error)
@@ -185,132 +189,6 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         }
 
         waitForExpectations(timeout: 1)
-    }
-
-    func testStartPayment_successfulResult_callsCompletionBlock() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-        mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
-            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
-            "assetsUrl": "http://assets.example.com"
-        ])
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
-        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
-
-        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
-        client.startPaymentFlow(threeDSecureRequest) { result, error in
-            XCTAssertNil(error)
-            XCTAssertNotNil(result)
-            paymentFinishedExpectation.fulfill()
-        }
-
-        threeDSecureRequest.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,%22description%22:%22ending+in+02%22,%22consumed%22:false,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true,%22status%22:%22authenticate_successful%22,%22enrolled%22:%22Y%22%7D,%22details%22:%7B%22lastTwo%22:%2202%22,%22lastFour%22:%220002%22,%22cardType%22:%22Visa%22%7D,%22bin_data%22:%7B%22prepaid%22:%22Unknown%22,%22healthcare%22:%22Unknown%22,%22debit%22:%22Unknown%22,%22durbin_regulated%22:%22Unknown%22,%22commercial%22:%22Unknown%22,%22payroll%22:%22Unknown%22,%22issuing_bank%22:%22Unknown%22,%22country_of_issuance%22:%22Unknown%22,%22product_id%22:%22Unknown%22%7D%7D,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true%7D,%22success%22:true%7D")!)
-
-        waitForExpectations(timeout: 2)
-    }
-
-    func testStartPayment_returnsFailedAuthenticationError_whenErrorReturnedInURL() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-        mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
-            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
-            "assetsUrl": "http://assets.example.com"
-        ])
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
-        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
-
-        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
-        client.startPaymentFlow(threeDSecureRequest) { result, error in
-            XCTAssertNotNil(error)
-            XCTAssertNil(result)
-            guard let error = error as NSError? else { return }
-            XCTAssertEqual(error.domain, BTThreeDSecureError.errorDomain)
-            XCTAssertEqual(error.code, BTThreeDSecureError.failedAuthentication("").errorCode)
-            paymentFinishedExpectation.fulfill()
-        }
-
-        threeDSecureRequest.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22threeDSecureInfo%22:%7B%22liabilityShifted%22:false,%22liabilityShiftPossible%22:true%7D,%22error%22:%7B%22message%22:%22Failed+to+authenticate,+please+try+a+different+form+of+payment.%22%7D,%22success%22:false%7D")!)
-
-        waitForExpectations(timeout: 2)
-    }
-
-    func testStartPayment_missingAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-        mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
-            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
-            "assetsUrl": "http://assets.example.com"
-        ])
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
-        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
-
-        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
-        client.startPaymentFlow(threeDSecureRequest) { result, error in
-            XCTAssertNotNil(error)
-            XCTAssertNil(result)
-            guard let error = error as NSError? else { return }
-            XCTAssertEqual(error.domain, BTThreeDSecureError.errorDomain)
-            XCTAssertEqual(error.code, BTThreeDSecureError.failedAuthentication("").errorCode)
-            XCTAssertEqual(error.localizedDescription, "Auth Response missing from URL.")
-            paymentFinishedExpectation.fulfill()
-        }
-
-        threeDSecureRequest.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?no-auth=bad-response")!)
-
-        waitForExpectations(timeout: 2)
-
-        XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.missing-auth-response"))
-    }
-
-    func testStartPayment_invalidAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-        mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
-            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
-            "assetsUrl": "http://assets.example.com"
-        ])
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
-        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
-
-        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
-        client.startPaymentFlow(threeDSecureRequest) { result, error in
-            XCTAssertNotNil(error)
-            XCTAssertNil(result)
-            guard let error = error as NSError? else { return }
-            XCTAssertEqual(error.domain, BTThreeDSecureError.errorDomain)
-            XCTAssertEqual(error.code, BTThreeDSecureError.failedAuthentication("").errorCode)
-            XCTAssertEqual(error.localizedDescription, "Auth Response JSON parsing error.")
-            paymentFinishedExpectation.fulfill()
-        }
-
-        threeDSecureRequest.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,BAD-JSON")!)
-
-        waitForExpectations(timeout: 2)
-
-        XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.invalid-auth-response"))
-    }
-
-    func testStartPayment_unexpectedAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-        mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
-            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
-            "assetsUrl": "http://assets.example.com"
-        ])
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
-        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
-
-        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
-        client.startPaymentFlow(threeDSecureRequest) { result, error in
-            XCTAssertNotNil(error)
-            XCTAssertNil(result)
-            guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTThreeDSecureError.errorDomain)
-            XCTAssertEqual(error.code, BTThreeDSecureError.failedAuthentication("").errorCode)
-            XCTAssertEqual(error.localizedDescription, "Auth Response JSON parsing error.")
-            paymentFinishedExpectation.fulfill()
-        }
-
-        threeDSecureRequest.handleOpen(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%22STRING%22")!)
-
-        waitForExpectations(timeout: 2)
-
-        XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.invalid-auth-response"))
     }
 
     func getAuthRequiredLookupResponse() -> [String : Any] {
@@ -388,7 +266,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
         mockAPIClient.cannedResponseBody = BTJSON(value: responseBody)
 
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
         client.startPaymentFlow(threeDSecureRequest) { _, _ in }
 
         waitForExpectations(timeout: 4)
@@ -400,7 +278,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
     }
 
     func testStartPayment_failure_sendsAnalyticsEvents() {
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
@@ -428,8 +306,8 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
         ])
-        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
-
+        let client = BTThreeDSecureClient(apiClient: mockAPIClient)
+        client.startPaymentFlow(threeDSecureRequest) {_,_ in } // TODO: - shouldn't have to do this for test to pass to inject a request
         let expectation = expectation(description: "willCallCompletion")
 
         threeDSecureRequest.nonce = "fake-card-nonce"
